requirements.txt
django-annoying